Refactoring of CLI script execution. Allowing multiple scripts to be executed in one command line.
