+++ TabularEditorTest/TestData/Silly02-WithErrors.csx
@@ -1,0 +1,1 @@
+﻿DOES NOT COMPILE
